refactor: publish core TX interlock contract (MOR-1697)
